requirements.txt
wheel
nbresult
colorama
ipdb
ipykernel
yapf
pygeohash
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.2.1
google-cloud-bigquery
google-cloud-storage
gcsfs
streamlit
db-dtypes
pyarrow

# prevent bq 3+ db-dtypes package error when running training from bq


python-dotenv
psycopg2-binary

# API
fastapi
pytz
uvicorn
# tests
httpx
pytest-asyncio
